Struct dioxus_signals::CopyValue

pub struct CopyValue<T: 'static, S: Storage<T> = UnsyncStorage> { /* private fields */ }

CopyValue is a wrapper around a value to make the value mutable and Copy.

It is internally backed by [generational_box::GenerationalBox].

Implementations

impl<T: 'static> CopyValue<T>

pub fn new(value: T) -> Self

Create a new CopyValue. The value will be stored in the current component.

Once the component this value is created in is dropped, the value will be dropped.

pub fn new_in_scope(value: T, scope: ScopeId) -> Self

Create a new CopyValue. The value will be stored in the given scope. When the specified scope is dropped, the value will be dropped.

impl<T: 'static, S: Storage<T>> CopyValue<T, S>

pub fn new_maybe_sync(value: T) -> Self

Create a new CopyValue. The value will be stored in the current component.

Once the component this value is created in is dropped, the value will be dropped.

pub fn new_maybe_sync_in_scope(value: T, scope: ScopeId) -> Self

Create a new CopyValue. The value will be stored in the given scope. When the specified scope is dropped, the value will be dropped.

pub fn take(&self) -> T

Take the value out of the CopyValue, invalidating the value in the process.

pub fn origin_scope(&self) -> ScopeId

Get the scope this value was created in.

pub fn id(&self) -> GenerationalBoxId

Get the generational id of the value.