Trait dioxus_signals::ReadableVecExt

pub trait ReadableVecExt<T: 'static>: Readable<Target = Vec<T>> {
    // Provided methods
    fn len(&self) -> usize { ... }
    fn is_empty(&self) -> bool { ... }
    fn first(&self) -> Option<ReadableRef<Self, T>> { ... }
    fn last(&self) -> Option<ReadableRef<Self, T>> { ... }
    fn get(&self, index: usize) -> Option<ReadableRef<Self, T>> { ... }
    fn iter(&self) -> ReadableValueIterator<'_, Self> 
       where Self: Sized { ... }
}

An extension trait for Readable<Vec> that provides some convenience methods.

Provided Methods

fn len(&self) -> usize

Returns the length of the inner vector.

fn is_empty(&self) -> bool

Returns true if the inner vector is empty.

fn first(&self) -> Option<ReadableRef<Self, T>>

Get the first element of the inner vector.

fn last(&self) -> Option<ReadableRef<Self, T>>

Get the last element of the inner vector.

fn get(&self, index: usize) -> Option<ReadableRef<Self, T>>

Get the element at the given index of the inner vector.

fn iter(&self) -> ReadableValueIterator<'_, Self>
where Self: Sized,

Get an iterator over the values of the inner vector.

Object Safety

This trait is not object safe.