dioxus_signals

Type Alias GlobalMemo

pub type GlobalMemo<T> = Global<Memo<T>, T>;

A memo that can be accessed from anywhere in the application and created in a static

Aliased Type

struct GlobalMemo<T> { /* private fields */ }

Implementations

impl<T: PartialEq + 'static> GlobalMemo<T>

pub fn id(&self) -> GenerationalBoxId

Get the generational id of the signal.

pub fn memo(&self) -> Memo<T>

Resolve the global memo. This will try to get the existing value from the current virtual dom, and if it doesn’t exist, it will create a new one.